+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 15
+
+* Render using multiple threads
+
 # 14
 
 * Update colouring